changed data_viz package structure
